Migrated sitemap and metatags API to App Router
