+++ app/(marketing)/@modal/(...)stats/[key]/modal.tsx
@@ -5,8 +5,7 @@ import { useRouter } from "next/navigation";
 import { AnimatePresence, motion } from "framer-motion";
 
 export default function Modal({ children }) {
-  const backdrop = useRef<HTMLDivElement | null>(null);
-  const wrapper = useRef<HTMLDivElement | null>(null);
+  const modalRef = useRef<HTMLDivElement | null>(null);
   const router = useRouter();
 
   const onKeyDown = useCallback((e: KeyboardEvent) => {
@@ -24,23 +23,26 @@ export default function Modal({ children }) {
     <AnimatePresence>
       <div className="absolute">
         <motion.div
-          ref={wrapper}
+          ref={modalRef}
           key="desktop-modal"
           initial={{ scale: 0.95 }}
           animate={{ scale: 1 }}
           exit={{ scale: 0.95 }}
           className="fixed inset-0 z-40 hidden min-h-screen items-center justify-center sm:flex"
+          onMouseDown={(e) => {
+            if (modalRef.current === e.target) {
+              router.back();
+            }
+          }}
         >
           {children}
         </motion.div>
         <motion.div
-          ref={backdrop}
           key="modal-backdrop"
           initial={{ opacity: 0 }}
           animate={{ opacity: 1 }}
           exit={{ opacity: 0 }}
           className="fixed inset-0 z-30 bg-gray-100 bg-opacity-10 backdrop-blur"
-          onClick={() => router.back()}
         />
       </div>
     </AnimatePresence>
